# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kaeman public API。除 /v1/auth/minecraft/challenge 与 /v1/services/bsi/gateway 外，所有端点都经过令牌桶限流，超限返回 429 并携带 Retry-After 响应头；意外错误统一返回 500 信封，data 中含 traceID。",
        "title": "kaeman-public-api",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/auth/check": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "auth"
                ],
                "summary": "查询 public api key 归属人绑定的 Minecraft 身份",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-router_handler_v1_auth_api_CheckData"
                        }
                    },
                    "401": {
                        "description": "Key required / Invalid public api key",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    }
                }
            }
        },
        "/v1/auth/minecraft/challenge": {
            "get": {
                "description": "生成一个 5 分钟内有效的 serverID 挑战，客户端需让玩家在 Mojang 会话服务器上把该 serverID join 到指定服务器，再调用 /verify 完成验证。挑战为一次性，verify 成功后立即失效。",
                "tags": [
                    "auth"
                ],
                "summary": "获取 Minecraft 会话验证挑战",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-router_handler_v1_auth_minecraft_api_Challenge"
                        },
                        "headers": {
                            "Cache-Control": {
                                "type": "string",
                                "description": "no-store"
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    }
                }
            }
        },
        "/v1/auth/minecraft/verify": {
            "post": {
                "description": "用 challenge 返回的 serverID 和玩家 UUID 调用，服务端校验玩家已在 Mojang 会话服务器 join 该 serverID。验证成功后签发 24 小时有效的 Minecraft 会话 JWT，并立即消费掉挑战。",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "验证 Minecraft 会话并换取会话 token",
                "parameters": [
                    {
                        "description": "挑战与玩家 UUID，均支持带/不带横杠",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router_handler_v1_auth_minecraft_api.VerifyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-router_handler_v1_auth_minecraft_api_VerifyData"
                        },
                        "headers": {
                            "Cache-Control": {
                                "type": "string",
                                "description": "no-store"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body / Invalid serverID / Invalid UUID",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired challenge / Minecraft session verification failed / Challenge already used or expired",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    }
                }
            }
        },
        "/v1/services/bsi/gateway": {
            "get": {
                "description": "建立 WebSocket 连接后作为只读订阅者接收服务端推送的 BSIMessage 文本帧（JSON）：type 为 chat 时 data 为 BSIChatResponse，type 为 splash 时 data 为 BSISplashResponse。客户端发送会被 CloseRead 忽略，服务端每 30 秒 ping 一次，空闲写超时 5 秒。OpenAPI 不支持 WebSocket，仅作说明用。",
                "tags": [
                    "services"
                ],
                "summary": "BSI 实时消息 WebSocket 网关",
                "responses": {
                    "101": {
                        "description": "Switching Protocols (WebSocket upgrade)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable（hub 已关闭）",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    }
                }
            }
        },
        "/v1/services/bsi/queue": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需要 splasher:queue 权限的 public api key，且 Authorization 中的 JWT 需同时携带 public API 与 Minecraft 会话 claims（audience 含 public-api 和 minecraft-session，data 同时包含 uid/permissions/ratelimit 与 minecraft）。消息经校验后分配无横杠 UUID 作为 id，实时广播给所有 /gateway WebSocket 订阅者。成功返回 202，data 为 null。",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "services"
                ],
                "summary": "向 BSI 广播一条消息",
                "parameters": [
                    {
                        "description": "type 为 chat 时 data 为 BSIChatRequest，type 为 splash 时 data 为 BSISplashRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router_handler_v1_services_bsi_api.BSIEnvelope"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "400": {
                        "description": "Invalid request body / Message must not be blank / Valid time, lobby and lobbyID are required / Unknown message kind",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "401": {
                        "description": "Key required / Invalid public api key / Required permission missing / Invalid minecraft key / Minecraft identity not bound to this public api key",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kaeman-dev_kaeman-public-api_response.Response-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_kaeman-dev_kaeman-public-api_response.Response-any": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 响应数据，失败或无数据时省略"
                },
                "msg": {
                    "description": "Msg 响应消息，通用消息为 HTTP 状态文本，业务消息为具体语义",
                    "type": "string"
                }
            }
        },
        "github_com_kaeman-dev_kaeman-public-api_response.Response-router_handler_v1_auth_api_CheckData": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 响应数据，失败或无数据时省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/router_handler_v1_auth_api.CheckData"
                        }
                    ]
                },
                "msg": {
                    "description": "Msg 响应消息，通用消息为 HTTP 状态文本，业务消息为具体语义",
                    "type": "string"
                }
            }
        },
        "github_com_kaeman-dev_kaeman-public-api_response.Response-router_handler_v1_auth_minecraft_api_Challenge": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 响应数据，失败或无数据时省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/router_handler_v1_auth_minecraft_api.Challenge"
                        }
                    ]
                },
                "msg": {
                    "description": "Msg 响应消息，通用消息为 HTTP 状态文本，业务消息为具体语义",
                    "type": "string"
                }
            }
        },
        "github_com_kaeman-dev_kaeman-public-api_response.Response-router_handler_v1_auth_minecraft_api_VerifyData": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data 响应数据，失败或无数据时省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/router_handler_v1_auth_minecraft_api.VerifyData"
                        }
                    ]
                },
                "msg": {
                    "description": "Msg 响应消息，通用消息为 HTTP 状态文本，业务消息为具体语义",
                    "type": "string"
                }
            }
        },
        "permission.Permission": {
            "type": "integer",
            "format": "int64",
            "enum": [
                1,
                2,
                4
            ],
            "x-enum-comments": {
                "BaseDefault": "base:default",
                "BaseDeveloper": "base:developer",
                "SplasherQueue": "splasher:queue"
            },
            "x-enum-descriptions": [
                "base:developer",
                "base:default",
                "splasher:queue"
            ],
            "x-enum-varnames": [
                "BaseDeveloper",
                "BaseDefault",
                "SplasherQueue"
            ]
        },
        "router_handler_v1_auth_api.CheckData": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "description": "ExpiresAt public api key 过期时间",
                    "type": "string",
                    "example": "2026-09-30T12:00:00Z"
                },
                "minecrafts": {
                    "description": "Minecrafts public api key 归属人绑定的全部 Minecraft 身份，小写无横杠 UUID",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "069a79f444e94726a5befca90e38aaf5"
                    ]
                },
                "permissions": {
                    "description": "Permissions public api key 的权限位集合，按位与判断",
                    "allOf": [
                        {
                            "$ref": "#/definitions/permission.Permission"
                        }
                    ],
                    "example": 4
                }
            }
        },
        "router_handler_v1_auth_minecraft_api.Challenge": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "description": "ExpiresAt 挑战过期时间",
                    "type": "string",
                    "example": "2026-09-29T12:05:00Z"
                },
                "serverID": {
                    "description": "ServerID 一次性挑战 ID，小写无横杠 UUID，签发后 5 分钟内有效且只能消费一次",
                    "type": "string",
                    "example": "0f8f8c8e5d3c4b2a9e7d6c5b4a3f2e1d"
                }
            }
        },
        "router_handler_v1_auth_minecraft_api.VerifyData": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "description": "AccessToken Minecraft 会话 JWT，audience 为 minecraft-session，24 小时有效",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOlsibWluZWNyYWZ0LXNlc3Npb24iXSwiZGF0YSI6eyJtaW5lY3JhZnQiOnsibmFtZSI6Ik5vdGNoIiwidXVpZCI6IjA2OWE3OWY0NDRlOTQ3MjZhNWJlZmNhOTBlMzhhYWY1In19fQ.x"
                },
                "expiresAt": {
                    "description": "ExpiresAt 会话 token 过期时间",
                    "type": "string",
                    "example": "2026-09-30T12:00:00Z"
                }
            }
        },
        "router_handler_v1_auth_minecraft_api.VerifyInput": {
            "type": "object",
            "required": [
                "serverID",
                "uuid"
            ],
            "properties": {
                "serverID": {
                    "description": "ServerID challenge 返回的一次性挑战 ID，有/无横杠均可",
                    "type": "string",
                    "example": "0f8f8c8e5d3c4b2a9e7d6c5b4a3f2e1d"
                },
                "uuid": {
                    "description": "UUID 玩家 Mojang UUID，有/无横杠均可",
                    "type": "string",
                    "example": "069a79f444e94726a5befca90e38aaf5"
                }
            }
        },
        "router_handler_v1_services_bsi_api.BSIEnvelope": {
            "type": "object"
        },
        "router_handler_v1_services_bsi_api.BSIType": {
            "type": "string",
            "enum": [
                "chat",
                "splash"
            ],
            "x-enum-varnames": [
                "BSITypeChat",
                "BSITypeSplash"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT。public API key 的 audience 为 public-api，携带 uid/permissions/ratelimit claims；Minecraft 会话 token 的 audience 为 minecraft-session，携带 minecraft identity claims。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}